requirements.txt
matplotlib==3.9.2
numpy==1.24.3
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
